+++ requirements.txt
@@ -1,8 +1,4 @@
 PyQt6
-random
-logging
-sys
-csv
-functools
-csv
-typing
+requests
+packaging
+pyqtdarktheme
